Avoid putting two single barlines at the start of a line.
This is a temporary fix. iReal allows a | at the end of a line
and also another | at the beginning of the next line.

Examples are in songs such as Ahmid-6, All Blues, Butterfly Dreams.

This needs to be revisited, as the ends of lines in the tune
Farmer's Trust are still not rendered correctly. A | at the end of
a line should always explicitly render a barline.