Reworked handling of barlines
The '|' in iReal serves two purposes: close the preceding bar and open the new bar.
It is also possible there will be one '|' to close a bar and another '|' to open
a new bar, possibly separated by other tokens such as Y or *B, etc.
In obj.bars, a single barline on the left will be '(', and on the right will be ')'.
The ')' is only used at the end of a row, but is always included in the obj.bars.
Instead of looking ahead for a closing bar, we now keep a reference to the previous
object in prevobj, and when encountering a closing bar, add it to prevobj.bars (and
reset prevobj to null so as not to add one twice).
For a '|', we both add ')' to prevobj and '(' to obj.
When rendering, we no longer need to check for an existing irr-rbar.
When adding an irr-rbar for Single-Barline, it must be inserted, not appended,
otherwise it floats right at the wrong height if preceded by an "over" bass note.

This change fixes unclosed lines in Farmers Trust from Jazz1350, as well as other
tunes with long measures at the end of a line.