Don't start new row on cell.spacer
When iReal has a Y spacer in the middle of a line, it still fills
the remaining cells. So if a new row is started when cell.spacer is
set, the cells all get out of alignment. This is shown in the tunes
Blue Daniel and Little Rootie Tootie from Jazz1350.

At the moment, a Y in the middle of a line will not be rendered.
TODO - render spacing when Y appears in the middle of a line.