Some more rendering fixes (#5)
* Fix missing alternate chords over p and space.

Used in the iReal song "At Long Last Love".

* Added missing "add" to the chord regex.

Used in the iReal songs "April Joy", "Dienda" and others.

* Avoid putting two single barlines at the start of a line.

This is a temporary fix. iReal allows a | at the end of a line
and also another | at the beginning of the next line.

Examples are in songs such as Ahmid-6, All Blues, Butterfly Dreams.

This needs to be revisited, as the ends of lines in the tune
Farmer's Trust are still not rendered correctly. A | at the end of
a line should always explicitly render a barline.

* Use proper flat and sharp in modifiers and title.

* Separate regex for real chords from regex for space, W and p.

Space, W and p must not be followed by modifiers or comment,
but can still have an alternate chord.
This fixes parsing confusion in New York, New York from Jazz1350.

* Reworked handling of barlines

The '|' in iReal serves two purposes: close the preceding bar and open the new bar.
It is also possible there will be one '|' to close a bar and another '|' to open
a new bar, possibly separated by other tokens such as Y or *B, etc.
In obj.bars, a single barline on the left will be '(', and on the right will be ')'.
The ')' is only used at the end of a row, but is always included in the obj.bars.
Instead of looking ahead for a closing bar, we now keep a reference to the previous
object in prevobj, and when encountering a closing bar, add it to prevobj.bars (and
reset prevobj to null so as not to add one twice).
For a '|', we both add ')' to prevobj and '(' to obj.
When rendering, we no longer need to check for an existing irr-rbar.
When adding an irr-rbar for Single-Barline, it must be inserted, not appended,
otherwise it floats right at the wrong height if preceded by an "over" bass note.

This change fixes unclosed lines in Farmers Trust from Jazz1350, as well as other
tunes with long measures at the end of a line.

* Don't start new row on cell.spacer

When iReal has a Y spacer in the middle of a line, it still fills
the remaining cells. So if a new row is started when cell.spacer is
set, the cells all get out of alignment. This is shown in the tunes
Blue Daniel and Little Rootie Tootie from Jazz1350.

At the moment, a Y in the middle of a line will not be rendered.
TODO - render spacing when Y appears in the middle of a line.